Adding docstrings to eval scripts. Fixing split import
