+++ python-sdk/nuscenes/eval/predict/do_inference.py
@@ -11,7 +11,7 @@ from nuscenes.eval.common.config import config_factory
 from nuscenes.eval.predict.data_classes import Prediction
 from nuscenes.predict import PredictHelper
 from nuscenes.predict.models import ConstantVelocityHeading
-from nuscenes.utils.splits import get_prediction_challenge_split
+from nuscenes.eval.predict.splits import get_prediction_challenge_split
 
 
 def do_inference_for_submission(helper: PredictHelper,
@@ -61,12 +61,12 @@ def main(version: str, data_root: str, split_name: str, output_dir: str, submiss
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description='Perform Inference with baseline models.')
-    parser.add_argument('-version', help='NuScenes version number.')
-    parser.add_argument('-data_root', help='Root directory for NuScenes json files.')
-    parser.add_argument('-split_name', help='Data split to run inference on.')
-    parser.add_argument('-output_dir', help='Directory to store output file.')
-    parser.add_argument('-submission_name', help='Name of the submission to use for the results file.')
-    parser.add_argument('-config_name', help='Name of the config file to use', default='predict_2020_icra')
+    parser.add_argument('--version', help='NuScenes version number.')
+    parser.add_argument('--data_root', help='Root directory for NuScenes json files.')
+    parser.add_argument('--split_name', help='Data split to run inference on.')
+    parser.add_argument('--output_dir', help='Directory to store output file.')
+    parser.add_argument('--submission_name', help='Name of the submission to use for the results file.')
+    parser.add_argument('--config_name', help='Name of the config file to use', default='predict_2020_icra')
 
     args = parser.parse_args()
     main(args.version, args.data_root, args.split_name, args.output_dir, args.submission_name, args.config_name)
